+++ broker-daemon/models/block-order.spec.js
@@ -880,93 +880,4 @@ describe('BlockOrder', () => {
       })
     })
   })
-
-  describe('::activeAmountsForOrders', () => {
-    const OrderStateMachine = BlockOrder.__get__('OrderStateMachine')
-    const { CREATED, PLACED, EXECUTING } = OrderStateMachine.STATES
-
-    it('only returns amounts for records in an active state', async () => {
-      const orders = [{ state: 'BAD_STATE' }]
-      const res = await BlockOrder.activeAmountsForOrders(orders)
-      expect(res.inbound.toString()).to.be.eql('0')
-      expect(res.outbound.toString()).to.be.eql('0')
-    })
-
-    it('returns amounts for created orders', async () => {
-      const orders = [{
-        state: CREATED,
-        order: {
-          inboundAmount: 100,
-          outboundAmount: 2000
-        }
-      }]
-      const res = await BlockOrder.activeAmountsForOrders(orders)
-      expect(res.inbound.toString()).to.be.eql('100')
-      expect(res.outbound.toString()).to.be.eql('2000')
-    })
-
-    it('returns amounts for placed orders', async () => {
-      const orders = [{
-        state: PLACED,
-        order: {
-          inboundAmount: 1000,
-          outboundAmount: 300
-        }
-      }]
-      const res = await BlockOrder.activeAmountsForOrders(orders)
-      expect(res.inbound.toString()).to.be.eql('1000')
-      expect(res.outbound.toString()).to.be.eql('300')
-    })
-
-    it('returns amounts for executing orders', async () => {
-      const orders = [{
-        state: EXECUTING,
-        order: {
-          inboundFillAmount: 10000,
-          outboundFillAmount: 50
-        }
-      }]
-      const res = await BlockOrder.activeAmountsForOrders(orders)
-      expect(res.inbound.toString()).to.be.eql('10000')
-      expect(res.outbound.toString()).to.be.eql('50')
-    })
-  })
-
-  describe('::activeAmountsForFills', () => {
-    const FillStateMachine = BlockOrder.__get__('FillStateMachine')
-    const { CREATED, FILLED } = FillStateMachine.STATES
-
-    it('only returns amounts for records in an active state', async () => {
-      const fills = [{ fill: { state: 'BAD_STATE' } }]
-      const res = await BlockOrder.activeAmountsForFills(fills)
-      expect(res.inbound.toString()).to.be.eql('0')
-      expect(res.outbound.toString()).to.be.eql('0')
-    })
-
-    it('returns amounts for created fills', async () => {
-      const fills = [{
-        fill: {
-          state: CREATED,
-          inboundAmount: 100,
-          outboundAmount: 2000
-        }
-      }]
-      const res = await BlockOrder.activeAmountsForFills(fills)
-      expect(res.inbound.toString()).to.be.eql('100')
-      expect(res.outbound.toString()).to.be.eql('2000')
-    })
-
-    it('returns amounts for a filled fill', async () => {
-      const fills = [{
-        fill: {
-          state: FILLED,
-          inboundAmount: 1000,
-          outboundAmount: 300
-        }
-      }]
-      const res = await BlockOrder.activeAmountsForFills(fills)
-      expect(res.inbound.toString()).to.be.eql('1000')
-      expect(res.outbound.toString()).to.be.eql('300')
-    })
-  })
 })
